added admin functions

+++ Models/Album.cs
@@ -2,7 +2,7 @@ namespace ITunesSearchApp.Models
 {
     public class Album
     {
-        public int CollectionId { get; set; }
+        public long CollectionId { get; set; }
         public string CollectionName { get; set; } = string.Empty;
         public string ArtistName { get; set; } = string.Empty;
         public string ArtworkUrl100 { get; set; } = string.Empty;
